==> code/1b-moisture_markdown.Rmd <==
---
output:
  github_document:
    html_preview: true
---

<!-- README.md is generated from README.Rmd. Please edit that file -->

```{r knitrsetup, echo=FALSE, message=FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  fig.path = "images_moisture_tracking/"
)
options(tibble.print_min = 5, tibble.print_max = 5)
source("code/0-hysteresis_packages.R")
theme_set(theme_bw())
```

# HYSTERESIS
Tracking moisture in soil cores

Run: `r Sys.Date()`

```{r, echo=FALSE, message=FALSE}
source("code/1-moisture_tracking.R")

mass %>% group_by(Core) %>% 
  dplyr::mutate(Hours = as.integer(difftime(Stop_datetime, min(Stop_datetime), units = "hours"))) ->
  massdata

ggplot(mass, aes(Stop_datetime, Moisture_perc)) +
  geom_point() + geom_line()+
  facet_wrap(~Core_assignment)+
  ylim(0,175)

massdata %>% 
  group_by(Core) %>% 
  mutate(moisture = if_else(Stop_datetime == max(Stop_datetime),Moisture_perc,as.numeric(NA))) %>% 
  dplyr::filter(!is.na(moisture)) %>% 
  ungroup %>% 
  group_by(Core_assignment) %>% 
  dplyr::summarize(moisture = round(mean(moisture),2),
                   time = max(Stop_datetime))->current_moisture

print("Current moisture conditions")
knitr::kable(current_moisture)

```
